correzione arrotondamento interi

+++ query1.sql
@@ -1,4 +1,4 @@
-select movies.director, COALESCE(min(gross-budget),-1) AS min, COALESCE(max(gross-budget),-1) AS max , COALESCE(TRUNC(avg(gross-budget),2),-1) AS avg 
+select movies.director, COALESCE(min(gross-budget),-1) AS min, COALESCE(max(gross-budget),-1) AS max , COALESCE(TRUNC(avg((gross-budget)::decimal),2),-1) AS avg 
 from (select director 
 from directors 
 where (yearofbirth - date_part('year', CURRENT_DATE) >= 50)) as oldDirector left join movies
